requirements.txt
nltk
PyPDF2
rake-nltk
xlrd
pillow
markdown
python-pptx
